+++ ScopusWp/main.py
@@ -5,7 +5,7 @@ import optparse
 # TODO: Make a fail proof for the citation comment wordpress not well formed
 # TODO: Make the new citations functionality with update date
 # TODO: Make update date into the main reference
-
+# TODO: Make error safe for http request pool exceptions wrapper method for requests
 
 
 def main():
@@ -43,6 +43,6 @@ if __name__ == '__main__':
     #main()
     controller = TopController()
     try:
-        controller.update_citations_website()
+        controller.populate_website()
     finally:
         controller.close()